requirements.txt
pkg_resources==0.0.0
PyOpenGL==3.1.10
PySide6==6.6.3.1
PySide6_Addons==6.6.3.1
PySide6_Essentials==6.6.3.1
shiboken6==6.6.3.1
